+++ dev/notes.txt
@@ -47,8 +47,46 @@
 📄In order to continue doing TDD, I should have a sign-in form, and set up tests for the database, form, & api
 📄I should wait on impementing mdb until I SEE that I need it.
 
+✨ ways to optimize learning that I should remember:
+  - take short, mind-wandering breaks apart from ultradian breaks
+  - use bedtime to estimate ultradian breaks for the day (TODAY: 09:00)
+  - angle screen upward to trigger alertness
+  - consciously try to blink less
+  - do productive breaks during ultradian breaks (raise adrenalin, then: NSDR, walk, talk to someone, etc.)
+  - implement perceptible, non-intrusive white noise during learning
+  - MINIMIZE DISTRACTIONS
+
+
 - ✅ issue a put request to the owners collection prior to the issue-tracker tests, to ensure a controlled environment.
-- ⚒️🎉 add put request code to make setup test pass
+- ✅ look at the role of the controller in mflix
+  📄 the MoviesController calls the appropriate DAO methods for a method (eg. apiGetMovies => MovieDAO.getMovies), formats the obtained data in the expected json return object, and responds with said object. It does NOT handle errors; simply calls await DAO.METHOD().
+- ✅ look at the role of router in mflix
+  📄 the router file creates a new router instance, and sets up controller methods for different paths. It then exports the router.
+- ✅ draw a diagram depicting how a put request from the test file is processed and results in a response -- include:
+  📄 the data that is being passed on in every step
+  📄 the way the data being passed on in every step 
+  📄 the test file
+  📄 the respective router
+  📄 the respective controller
+  📄 the respective controller method
+  📄 the respective dao
+  📄 the respective dao method
+  📄 the respective db method
+- 🎉🧱 code everything leading UP to the OwnersDAO put issues method:
+  - ✅ create an issue-handler file in the controller folder
+  - 🎉 create an IssueHandler class in the owners-handler file
+  - create a putDocumentRequest method in the IssueHandler class, with params (req, res, next)
+  - export the IssueHandler class
+  - import the IssueHandler class in the api file
+  - map issues/:owner/:project put requests to the IssueHandler class' putDocumentRequest method
+  - log a message in the putDocumentRequest method of the OwnersHandler class
+  - verify said msg is being logged; if not, debug
+  - add documentation to the putDocumentRequest method
+    📄 simply describe params, and write that the method handles the put request that is submitted by the test file by calling the dao object's putDocument method. Will edit the method's response, but for now will simply respond with whatever the dao method returns.
+  - import the OwnersDAO object if it is not yet imported
+  - call the OwnersDAO putDocument method, with the document passed, and store the return value in a response variable
+  - return said response variable
+- 🧱 add put request code to the OwnersDAO to make setup test pass:
   - ✅ find a way to connect to db BEFORE the router method setups: connect in server?
   - ✅ log a message with put requests
   - ✅ check how the data is being sent
@@ -57,7 +95,7 @@
     📄 they use OOP to format apiCalls as methods, and use a different class called DAO (direct access object) which are the objects that access the db
   - ✅ look at how the DAO are connected to the index file
   - ✅ implement the mflix api db code
-  - code the ownersDAO get issues method
+  - 
 
 - ⚒️Create all of the functional tests in tests/issue-tracker.test.js:
   - Create an issue with every field: POST request to /api/issues/{project}
