+++ tests/issue-tracker.test.js
@@ -5,40 +5,35 @@ const chaiHttp = require('chai-http'),
   chai = require('chai'),
   server = require('../server.js'),
   { assert } = chai,
-  ISSUES_PATH = '/api/issues/solarc117/issue-tracker-test'
+  TEST_PATH = '/api/issues/test%20owner/test%20project'
 
 chai.use(chaiHttp)
 
 suite('🧪 Issue Tracker', () => {
   suiteSetup(done => {
-    chai
-      .request(server)
-      .put(ISSUES_PATH)
-      .send({
-        _id: new ObjectId(1),
-        owner: 'solarc117',
-        projects: [
+    const testDocument = {
+      _id: new ObjectId('0000000197d9af3844c5dc90'),
+      owner: 'test owner',
+      project_issues: {
+        'test project': [
           {
-            name: 'issue-tracker-test',
-            issues: [
-              {
-                title: 'test-issue',
-                text: 'this is a test issue',
-                created_by: 'solarc117',
-                assigned_to: null,
-                status_text: null,
-              },
-            ],
+            title: 'test issue',
+            text: 'this is a test issue',
+            created_by: 'test owner',
+            assigned_to: null,
+            status_text: null,
           },
         ],
-      })
-      .end(done)
+      },
+    }
+
+    chai.request(server).put(TEST_PATH).send(testDocument).end(done)
   })
 
   test('Correct setup document', done => {
     chai
       .request(server)
-      .get(ISSUES_PATH)
+      .get(`${TEST_PATH}`)
       .end((err, res) => {
         const {
           status,
@@ -58,7 +53,7 @@ suite('🧪 Issue Tracker', () => {
   test('1. GET /api/issues/:owner/:project', done => {
     chai
       .request(server)
-      .get(ISSUES_PATH)
+      .get(TEST_PATH)
       .end((err, res) => {
         const { status, ok, body } = res
